feat: exception

+++ blog-inf/src/main/java/com/star/inf/dto/PageData.java
@@ -31,7 +31,7 @@ public class PageData<T> {
     /**
      * 总数
      */
-    @ApiModelProperty(name = "count", value = "总数", required = true, dataType = "Integer")
+    @ApiModelProperty(name = "count", value = "总数", required = true, dataType = "Long")
     private Long count;
 
 }
